[COR-3748] Updated to use V2507
